systemd service script for piTelex
This script allows starting piTelex as a service (installation see
README). piTelex is started inside a byobu session to facilitate
headless operation: Just log in via SSH and enter byobu-screen to attach
to the piTelex instance, detach with F6.

NB: You need to edit service file for piTelex path and user (default is
pi, not root!).