txDevED1000SC, txConfig: include recv_debug in defaults
Option recv_debug was up until now not included in the default
ED1000SC configuration, and defaulted to on. That meant quite a disk
space hog for longer runtimes, as the file is always appended per
default. This also puts strain on the flash memory when used in a
Raspberry Pi.

Now include in default configuration file, and default to off. Most
users won't know how to interpret this file and probably won't need to.