txDevITelexSrv: Bind to server socket despite TIME_WAIT connections
Problem was: After quitting telex.py, some connection remnants lingered
(namely self-tests) and prevented a new telex.py instance from binding
to the server socket.

This is normal TCP behaviour and is designed to, after a service exits,
prevent delayed packets intended for it reaching a new, succeeding
"owner" of the socket. The connection is held in TIME_WAIT (or
CLOSE_WAIT, depending on who terminates the connection) state for 2*MSL
(2*120 s as per standard).

In this case, especially regarding the self-test package, no ill
side-effects are to be expected. Hence set the socket's SO_REUSEADDR
option to enable binding nonetheless.