txDevITelexClient: Fix bug in keyboard dial
There was a bug with different dialling methods reproducible like so:

Capture packets of two connections (e.g. using wireshark with capture
filter "tcp port <target's i-Telex port>") using different dialling
methods:

1. Pressing <ESC> # <n1> <n2> ... <Enter> establishes connection
   immediately.

2. Pressing <ESC> <Insert> <Enter> triggers WB (ready to dial), then
   dial a number. Connection is established on last digit.

With method 1, everything works as expected. Connection is established
to the proper endpoint.

Method 2 also connects properly, but in addition sends the dialled
number immediately to the recipient as soon as connection is
established. This may be ignored by "slow" hardware, but was observed to
interfere with an automatic telex service (727272 wetter d). Depending
on timing and chance, the digits were interpreted as input to the menu
system.

Interestingly, there exists a commented-out section in txDevITelexCommon
which corrects this issue. It was introduced in a major rewrite of
connection logic (d41d3b6). So uncomment it.