txDevITelex*: Raise self-test timeout to 3 s
During testing, some timeout events occurred with 1 s timeout. Ideally,
not the connection itself should time out, but the waited-on event for
confirmation by the server thread. Only the latter is being recorded in
the self-test logic; a connection timeout will only be logged and not
acted upon. Both had the same times set. Raise connection timeout to
have self-test fail as intended.