txDevITelexSrv.py: Handle ConnectionAbortedError
In rare corner cases, a client can successfully connect to us, but
instantly terminate the connection using RST. If the RST packet arrives
before calling accept, ConnectionAbortedError is raised, which, if
unhandled, terminates the server thread, rendering piTelex inoperable.

So catch the exception. As per Stevens, UNIX network programming, it's
safe to ignore the error and continue listening, and accepting the next
contestant.