txDevED1000SC: rx thread performance optimisation
It was apparent that with the ED1000SC module enabled, piTelex caused a
constant CPU load of 30...40% on a Raspberry Pi 4B. Using yappi, the
ED1000 rx thread could be identified as the culprit. This is
comprehensible since a software IIR filter is employed for demodulation
-- every 5 ms (four times per symbol), even in quiescent state. So the
following is changed:

- rx thread: If offline, check every 2 s for bit change (i.e. wait for
  AT pressed). The filter seems tuned in a way that it detects a level
  change after two scans. The worst case delay from button press to
  machine turn-on should now be 4 s. If online state is set by a call,
  reaction is immediate (see Event.wait).

  Also, don't wait 20 "Z" level scans before reacting to AT, but react
  on the first.

  Settings have been tested with Siemens T1000 and it performs
  satisfactorily; YMMV regarding the AT reaction time.

- tx thread: As a bonus, if there's nothing to do, don't poll every 100
  ms, but wait for online state. Probably no significant performance
  impact.

With the optimisation, CPU load drops to about 7% when printer is off.

Sample cpufreq statistics and temperature (vcgencmd measure_temp) after
about a day each:

- before: 600 MHz:0,00%, 750 MHz:93,19%, 1000 MHz:5,85%, 1.50 GHz:0,96%
  	  47°C

- after:  600 MHz:94.76%, 750 MHz:3.22%, 1000 MHz:1.02%, 1.50 GHz:0.99%
          44°C