txDevED1000SC: Allow tx buffer flush after going offline
Up until now, when the MCP wrote a ESC-Z command for going offline, we
obeyed instantly. On normal connections there was no problem. Automatic
services like the teleprinter.net news service though tend to send
faster than 50 Bd; so if they terminated the connection all as of yet
unprinted characters were lost.

To counter this, we

- don't empty the tx buffer when going offline and
- continue printing the buffer contents even if we're offline.

As soon as the buffer is empty, we're "really" offline.