txDevED1000SC: Drop "ready to dial" signal
In the piTelex documentation there is a diagram which shows how pressing
AT is handled on ED1000. The sequence is as follows:

- Both A and B on A level
- Press AT: B switches from A to Z level
- A raises to Z also
- After a while, A sends a ready to dial signal by pulsing Z-A-Z for
  40 ms.

On a Siemens T1000 which worked fine on i-Telex, on piTelex it always
printed a "v" after pressing AT. As a matter of fact, sending a V in
ITA2 is:

AAZZZZZz
SCCCCCTt

S: start bit
C: character bits, "v" being AZZZZ
Tt: stop bits (1.5)

With a bit time of 20 ms, this is a 40 ms space impulse, exactly like
the ready-to-dial signal above.

So remove the rtd pulse. For good measure, make the ready-to-dial signal
identical to incoming-call signal, involving 120 ms delay. Testing
with the above T1000 now showed identical function to i-Telex.

Legend:
- Stations: A: exchange, B: teletypewriter
- Levels: A: space, Z: mark