txDevED1000SC: Don't add start/stop bits when not sending character
The tx logic pops an item off the buffer and then decides if it's a
normal character or a special command. Only characters should have start
and stop bits added; the special commands create specially-tailored
pulses and should thus be exempt.

So only add start and stop bits if we send a character.