i-Telex-conformant welcome banner printing
According to the current specs and personal correspondence with Fred,
i-Telex welcome banner sending rules are as follows:

- Don't send banner if no printable data has been received, for
  "maintenance connections".

- Print a banner if printable data has been received, i.e. either Baudot
  or ASCII data. The banner must be printed before anything else.

To elaborate, a banner should be printed as soon as the printer has been
started, which should happen in the aforementioned second case only.

The previous logic would send a banner on the first socket timeout. On
manual connections, this would normally be indistinguishable from the
above requirement. But with automated ASCII connections that immediately
send data, a banner was mostly not printed at all, but sometimes (packet
loss) in the middle of received text.

New logic: Track connection state in four phases.

- disconnected
- connected, but printer not yet started
- connected, printer has been started, welcome banner hasn't been
  fully received yet
- connected, welcome banner has been received completely

Difficulty arises from the signal flow: If a banner should be printed,
only a command is sent to MCP, which generates and sends the banner
later. To ensure that the welcome banner is always printed first,
read()s from ITelexSrv are filtered in in state 2 so that only commands
will be read; printable data is hidden. MCP eventually receives the
welcome banner command and sends the banner plus a new special command
afterwards, which ends up in our tx buffer, to be printed. After
receiving the special end-of-welcome command, we enter state 3, and all
remaining data can be read() and processed as usual.

Further remarks:

- Incorporate the banner in received_count, to signal the other party
  properly how much there is still to print. To facilitate this, let
  send_banner return the (luckily) fixed length.

- Mute the special command in txDevScreen and txDevLog.

- Technically, we're still not up to the specs, since the banner should
  only be sent if the printer has in fact been started. Instead, we only
  queue the "start printer" command and the banner afterwards in the
  receive buffer. This should suffice for now, since this check would make
  piTelex unusable without a connected teletypewriter.