txDevED1000SC: No character recognition if offline
When teletypewriter was offline and switched to online by us, it raised
its send level from A to Z. Depending on timing, this triggered
reception of any character that begins with consecutive As, followed by
consecutive Zs (see comment in this commit).

A similar issue had been present when we were going online because we
detected that the AT is pressed on the teletypewriter. This commit
unifies the solution for both ways of going online (AT and incoming
connection).

(Also, change btype argument of signal.iirfilter to "bandpass" to make it
unambiguous; no functional change.)