telex.py: Create global error log
Thus far, exceptions were caught in a nonstandardised way and at least
some exceptions' names were output onto the telex console (those caught
and LOG()ged manually). This is nice for console users. But only if the
exception is caught and logged properly, and if the debug level is set
properly in txConfig.json.

So we set up a global exception log to register all exceptions for later
perusal. Everything is logged into a file rotated monthly.

Direct logging from modules is also possible and will be set up in the
next commit.

Also, clean up unused import.