txDevED1000SC: Fix rogue ST press after startup
After starting piTelex, thus far, the receiving bit detection logic
started completely blank. The consequence was that after 100 scans of
"0" input, an ST press was detected. This was less of a problem until
the offline scan rate was lowered to save cpu cycles; before, 100 scans
took 500 ms. After the aforementioned change, this was extended to 200 s
and could easily lead to an accidentally terminated connection.

To resolve this, on start, set the bit counter for "0" to 100 as if the
ST detection had just taken place. We can be quite sure that the
teleprinter is offline because

- we just started and
- we didn't receive a 1 (if we did, the zero counter will be reset).