txDevITelex*: Act on printer startup timeout
Previously, a printer feedback timer was introduced to txDevMCP. Modify
txDevITelex* to act on printer startup timeout (send "der" reject
packet). This only concerns us when we are server (i.e. being called).
Otherwise, if the printer is operated manually and doesn't start up, we
cannot print anything, so the failure will be noticed and we cannot
sensibly print any error.