txDevMCP: WRU fallback mode
Previously, if wru_id was set, txDevMCP would always respond on WRU.
Normally it would not be set if real hardware is connected. If the
printer didn't start up on an incoming call, the connection was
terminated.

Introduce new WRU fallback mode: If enabled,

- intercept ESC-ACT (printer start timeout) and replace by ESC-ACK
  (printer start confirmed) to prevent txDevITelexSrv from terminating
  the incoming connection, and
- enable software responder in this case.

This allows for recording messages transparently in case of printer
malfunction.