txDevED1000SC: Quick scanning on Z when offline
In a previous commit, receive scan rate was lowered substantially to
save CPU cycles. The disadvantage was longer reaction time on AT press
and, because we reacted on the first detected Z level alone, a certain
flimsiness to external signal disturbances (e.g. plugging the
teleprinter out or in).

To alleviate this, introduce quick scanning: At the first detection of Z
level, enable quick scanning, which means the _is_online.wait delay is
skipped. From here on, there are two possibilities:
- Z is stable for 20 scans (100 ms): Go online.
- An A is read: Disable quick scanning, bit counters are reset as usual.

With this modification, we get the robustness of the original 100 ms
debounce timing and also shorten our AT reaction time.

(Reaction time is actually shortened because the receive filter
typically needs two scans to recognise the A=>Z transition, which took
us 2 s before. Now, it's 1005 ms, worst case.)