txDevITelexCommon: Don't count welcome banner for Ack
Previously, on being connected as server, the first receive_count sent
was 24 (the welcome banner length). This lead to a deadlock as on the
other side, if it also was piTelex, because the unprinted counter
overran to 232. Sending data wasn't possible.

This might rather be a bug in the unprinted calculation, but this kind
of Acknowledge usage wasn't intended in the i-Telex protocol -- the
first ACK should be sent with 0. So fix the problem on this side, and
add a comment for a later fix.