txDevITelexCommon: Improvement in version packet handling
Previously, when we were client and receiving a version packet, we would
always send one of our own. Some services (like conference call service
v1.0.8.0, Tx 11160) would in turn send another version packet when
receiving one.

Actually, this is exactly what the currents specs say, but leads to an
infinite loop. In the above case, we would only send version packets but
no upstream data.

So fix this by introducing a field to save the remote's protocol
version. The new logic works as follows.

When receiving a version packet, save the version advertised therein.
Then send own version packet if we haven't yet sent one (always with
version 1). As client, we always send version as the first packet after
connecting.

If both versions are equal, there is nothing else to do.

If the received version packet contained a version different from 1, our
sending of a v1 packet is a request to, in turn, send a v1 packet to
signal that it's supported at the remote end.

So the next version packet is checked and acted upon. If the version is
different, reject; if not, proceed and ignore future version packets.