txDevMCP: Re-implement classic best-effort dialling
The scope of the past dial logic rework was i-Telex compatibility and
keyboard dialling convenience. As it turned out, the classic piTelex
dialling was *very* convenient for rotary dial users.

So reimplement it, selectable by setting the configuration variable
dial_timeout to 0 (previously forbidden). It's now called "instant
dialling".

A new ESC command had to be introduced because the current ESC-# prints
an error message and stops the teleprinter if an invalid number was
passed to it. Sending ESC-#!<number> tries dialling the number and fails
silently, which is not strictly i-Telex compliant, but very convenient
for some users.