telex.py: Make error log location configurable
To support setups where data and executables are stored seperately,
introduce config option to allow changing the error log path.

The error log base directory can be set in errlog_path configuration
option as

- relative path (interpreted relative to the location of piTelex files),
  or as
- absolute path.

Downside: Error log can only be initialised after loading the
configuration. This means all errors during config parsing and
processing aren't logged and are output to console only. Maybe there is
a solution to this, but it's not obvious. The risks should be marginal,
as configuration errors should be apparent when trying to start piTelex
manually.