txDevED1000SC: Make online delay configurable
Previously, when going online, the teleprinter had a maximum timeframe
of 500 ms (100 slices). In some use cases, this may not be enough, e.g.
if the teleprinter is turned on using a relay when going online. If this
threshold is exceeded, the printer is detected as unresponsive and
turned off again.

The following response times could be obtained with a Siemens T1000
using the txDevRelayPlate module:

- A=>Z transition when powered:              47 slices ( 235 ms)
- A=>Z transition after turning mains on:   473 slices (2365 ms)

So make the maximum allowable reaction time configurable
(unres_threshold). Also, introduce reporting log commands to facilitate
recording the reaction time.

In the above case, an unres_threshold of 600 slices (3 s) was selected
and yielded satisfying performance.

In the future, we might improve upon this by applying different
thresholds depending on the relay state (which is transparent
currently). The gain is questionable however, since the worst case -- a
truly unresponsive printer which is detected only a few seconds later --
will probably hardly lead to problems. Also, module interdependence
would rise, which will raise complexity and may result in unforeseen
problems.