txDevCH340: Send buffer feedback also when offline
Previously, printer buffer feedback was only sent when online (i.e., MCP
in WB/A mode). After termination of an i-Telex connection however, the
teleprinter typically is still emptying the buffer (and thus running),
but already set to offline (by ESC-Z). Feedback needs to be sent until
the buffer is empty to retrigger the power watchdog, preventing
premature power-off.