txDevMCP: Fix if/iteration bug leading to immediate power-off
There were if statements iterating over tuples, which had been changed
to seemingly single-element tuples. Because the trailing comma was
missing, the string inside was iterated over, leading to wrong matches.

Specifically:

if a in ("Z", "ZZ"):
    do thing

was changed to

 if a in ("Z"):
     do thing 1
 if a in ("ZZ"):
     do thing 2

If a equalled "Z", both ifs matched, the second one being equal to

 if a in ["Z", "Z"]:

Change to explicit comparison using ==, although as an alternative the
following may be used, explicitly creating single-element tuples:

 if a in ("Z",):
     do thing 1
 if a in ("ZZ",):
     do thing 2