txDevITelexCommon: Remote printer buffer feedback
Send remote printer buffer feedback as ESC-° with the number of
unprinted characters, like ESC-~ for locally unprinted characters.

This is a lot simpler than ESC-~ since we just trust the remote's
judgment about how often to send an i-Telex Acknowledge packet. On
receipt of such a packet, we calculate the number of remotely unprinted
characters and forward them inside ESC-°.

The principal reason for introducing this is the opposite case from the
reason for introducing ESC-~. This time, we send a lot quicker than the
receiver is able to print. While the remote teleprinter is printing, we
are idle for extended amounts of time, probably triggering the "active"
timeout and terminating the connection. ESC-° alleviates this by
signalling remote printing activity and this way we can reset our local
timer, preventing timeout.