Move find_rev to telex.py, clarify error msg, suppress stderr
Previously, find_rev was resident in txDevLog.py, but also used in
telex.py. This lead to confusion since txDevLog code was used even
though txDevLog was disabled in the configuration.

So move find_rev to telex.py and import in txDevLog.

Also, clarify error displayed on piTelex start (in error log) in case
the current git HEAD revision cannot be found.

Finally, suppress stderr so that, in case of error, the console isn't
being spammed, the rationale behind this being that the non-git case
will probably become normal (i.e. the normal user downloading piTelex as
zip file).