txDevITelexCommon: Properly print reason on Reject
The reason in a Reject package is given as ASCII characters, which we
receive as byte objects. Those bytes were passed to txCode's translate
function, which discarded them because it expects Python strings. This
entailed the teleprinter not printing the reason (e.g. occ) and the
error log entry stating an empty reason.

Fix by properly decoding received ASCII bytes to string before calling
translate.

Also, remove \x00 padding that may be introduced by i-Telex, e.g. when
sending nc as "nc\x00".

Additionally, add docstring to translate function to alert the user to
the necessity of converting received bytes to strings.