txDev{ED1000SC,ITelexCommon}: Ensure Python 3.5 compatibility
Revert enum.auto() usage, which is a convenience function to
automagically assign an integer value to states, to ensure Python 3.5
compatibility. It's been declared EOL on 5 September 2020, but still in
use in a lot of older Raspbian images.

Also, remove comments from txDevED1000SC that refer to obsolete state
values.