txDevITelexCommon: Connect outgoing ASCII immediately
As of yet, standard piTelex policy for all types of outgoing (=client)
connections was to only switch from WB (ready-to-dial) to A (connected)
if any data was received. [This does not concern incoming (=server)
connections because they don't employ the WB state.]

For i-Telex binary connections, this was sufficient (and required as per
spec) due to the mandatory i-Telex handshake and keep-alive packets
being exchanged automatically, and due to the fact that there are
i-Telex packets that are only of administrative use and should
explicitly not turn on the teleprinter.

For ASCII connections, there are no prescribed automatic administrative
packets like for the binary protocol. This means if an ASCII peer was
dialled and it didn't send any initial data, piTelex would never switch
from WB to A, preventing us in turn from sending anything to "trigger"
the peer to send anything back.

Fix this by introducing a special case into connection handling code for
ASCII client connections, immediately issuing ESC-A as soon as the TCP
connection is established. Also, insert a comment at the "normal" ASCII
processing code.

Editor's note: This was the case with the Minitelex Tx-to-Fax gateway
service, which behaves just like a Telex endpoint, i.e. it will only
ever send anything if its WRU unit is triggered.