txDevArchive: Ignore rogue ESC-WB commands
The archive module starts recording after ESC-A command and saves the
message at ESC-Z. The ESC-WB command is used to determine the outgoing
number (used for naming the respective archive file).

The typical sequences are:

1. ESC-A, message, ESC-Z (incoming)
2. ESC-WB, dialling, ESC-A, message, ESC-Z (outgoing)

If there happens to come an ESC-WB after ESC-A (currently this happens
with ED1000 if the TP has an error), it would reset the archive module
and prevent it from saving the current message. Catch this by ignoring
ESC-WB after having seen ESC-A.