txDevITelexCommon.py: Start motor when receiving Acknowledge
On incoming connections, the spec dictates that our peer send a Direct
Dial, after which we must start out motor. So far so good.

On outgoing connections, we trusted that our peer sent the date/time
banner as soon as its motor started successfully. So we started our
motor on receipt of Baudot Data.

The banner is not mandatory though, and can be disabled in i-Telex
configuration. If disabled, the peer only sends Acknowledge, which also
transports the information "remote motor started".

So modify Acknowledge code to start motor if not yet running. Copy the
relevant section from Direct Dial / Baudot Data.