txDevED1000SC: Change condition for sending printer buffer feedback
Printer buffer feedback triggers recalculation of the Acknowledge
counter in the i-Telex module, which is important for the sender to know
how much of the sent data we've printed yet.

Up until now, ESC-~ was sent only if the current buffer length changed.
When transmitting "full throttle" this is no problem, but when
transmitting only single characters this may cause the connection to
stall: The probability is high that single characters inside the print
buffer aren't noticed, so it's always 0, ESC-~ isn't sent and the
Acknowledge counter isn't updated and our peer may subsequently refuse
to send more data.

Change the logic to always send ESC-~ if the number of printed
characters has changed.