txDevRPiTTY.py: Ensure sending ESC-~0 on shutdown
txDevITelex* needs to know the buffer is empty to completely reset its
state (and thus make us reachable again via i-Telex), so it is
imperative that the hardware's last buffer feedback be 0 (empty).
Ensure this by recalculating the buffer independently from our state, so
that even after receipt of ESC-Z, while the rest of the buffer contents
(if present) are being printed, buffer feedback (ESC-~) will be sent
until we're finished.