txDevRPiCtrl: Add delay option for AT and ST inputs
Add an optional delay to make reactions to pressing AT and ST buttons
more realistic.

Thanks to Rolf for this patch.

NB: The button callbacks are called from a different thread (the one
communicating with pigpiod). Care needs to be taken when implementing
more elaborate features. For now, this should be safe without extra
synchronisation: Watchdog.restart, in this setup, only modifies one dict
entry. Worst case, timing may be off for one call of idle20Hz (meaning a
timing error of 50 ms).