txDevArchive.py: Critical: Remove bad chars in filename
The Archive module's filename method didn't check the wru argument for
bad characters. If there was a "/" inside the wru, it would carry over
to the filename. On all supported platforms however slashes are
forbidden.

Also, catch exceptions whilst writing the file. In one occasion, the
above happened (slash inside filename) and a FileNotFoundError was
raised and terminated the main loop.