txDevRPiTTY: Set "motor running" mode after pulse dial
When using the standard electronics for TW39 teleprinters with FSG,
there is a problem in recognising if the teleprinter has started. The
following descriptions reference piTelex standard electronics.

On incoming calls, everything is in order: The electronics reverses loop
polarity, which signals the FSG to turn on the teleprinter motor relay.
After turning on the relay, the FSG raises loop current from 5 mA to 40
mA, confirming teleprinter start-up. This current is detected by an
optocoupler inside the electronics, which feeds the line observer and
confirms start-up and causes us to send ESC-AA in turn.

On outgoing calls with number switch, the situation is different.
Pressing the AT causes the FSG to raise loop current immediately, which
signals the dialling whish to the electronics (entering WB mode). The
line observer must not take this as confirmation of start-up, since it
didn't occur yet. After dialling, when connection is established, we
receive ESC-A. Line polarity is reversed by the electronics, which
causes the FSG to turn on the motor relay. As the loop current is
already at 40 mA however, the optocoupler cannot determine if the
start-up has occurred. The dialling user can be sure of it however
because they observe the motor turning on. This is why we must not wait
for printer start confirmation in this case. The same applies to
keyboard dialling, albeit the motor is turned on earlier, when entering
WB.

This patch effects this behaviour by tracking state change. If we enter
S_ACTIVE_INIT after a dialling state, startup feedback is always
emitted.

Also, remove a redundant pass and correct typos.