txDevRPiTTY: Ensure to always send printer buffer updates
When sending from a 50 Bd to a 200 Bd machine, due to timing issues, the
receiving piTelex ceased to send buffer updates (ESC-~) and thus sent no
i-Telex Acknowledge telegrams to the sending piTelex. The latter then
ceased to send any data, believing the buffer was full.

Correct this to always send buffer updates as long as the teleprinter is
online.